Moved all state logic into StateMachines. Added TimerStateMachine. Prevented unnecessary component renders.
