+++ package.json
@@ -2,6 +2,7 @@
   "name": "metronome-prototype",
   "private": true,
   "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
