+++ src/config/commonInterfaces.ts
@@ -1,0 +1,5 @@
+export enum EStep {
+	PAUSE = "PAUSE",
+	LOW = "LOW",
+	HIGH = "HIGH",
+}
