--- src/hooks/useTimer/useTimerConsts.ts
@@ -1,1 +1,0 @@
-export const INITIAL_STEP_INDEX = 0;
