+++ src/hooks/useTimer_remove/useTimerConsts.ts
@@ -1,0 +1,1 @@
+// export const INITIAL_STEP_INDEX = 0;
