+++ src/hooks/useTimer_remove/useTimerInterfaces.ts
@@ -1,0 +1,8 @@
+// export type onTickHandler = (step: EStep) => void;
+
+// export interface IUseTimerProps {
+// 	pattern: EStep[];
+// 	tempo: number;
+// 	isPlaying: boolean;
+// 	onTickHandlers?: onTickHandler[];
+// }
