+++ src/services/MetronomeStateMachine/actors/Timer.ts
@@ -1,0 +1,22 @@
+// import { fromCallback, setup } from "xstate";
+// import { TimerCallback } from "@services/MetronomeStateMachine/actors/TimerConsts";
+
+// export const TimerActor = fromCallback(TimerCallback);
+// const TimerCallback = ;
+//
+// const resizeLogic = fromCallback(({ sendBack, receive }) => {
+// 	// Cleanup function
+// 	return () => {
+// 		console.log("Cleaning up");
+// 	};
+// });
+//
+// const myActorLogic = fromCallback(({ sendBack, receive }) => {});
+//
+// const machine = setup({
+// 	actors: {
+// 		resizeLogic,
+// 	},
+// });
+//
+//
