+++ src/services/MetronomeStateMachine/actors/TimerInterfaces.ts
@@ -1,0 +1,4 @@
+// export enum ETimerActions {
+// 	START = "start",
+// 	STOP = "stop",
+// }
