+++ src/services/MetronomeStateMachine/actors/TimerStateMachineActor/TimerStateMachineActor.ts
@@ -1,0 +1,11 @@
+import { ActorLogic, Snapshot } from "xstate";
+import {
+	ITimerStateMachineContext,
+	TTimerStateMachineEvent,
+} from "@services/MetronomeStateMachine/machines/TimerStateMachine/TimerStateMachineInterfaces";
+
+export interface ITimerStateMachineActorLogic
+	extends ActorLogic<
+		Snapshot<ITimerStateMachineContext>,
+		TTimerStateMachineEvent
+	> {}
