+++ src/services/MetronomeStateMachine/actors/TimerStateMachineActor/TimerStateMachineActorConsts.ts
@@ -1,0 +1,8 @@
+// import { createActor } from "xstate";
+// import { TimerStateMachine } from "@services/MetronomeStateMachine/machines/TimerStateMachine/TimerStateMachine";
+//
+// export const TimerStateMachineActor = createActor(TimerStateMachine, {
+// 	input: {
+// 		defaultRating: 3,
+// 	},
+// });
