+++ src/services/MetronomeStateMachine/machines/MetronomeStateMachine/MetronomeStateMachineConsts.ts
@@ -1,0 +1,22 @@
+import { IMetronomeContext } from "@services/MetronomeStateMachine/machines/MetronomeStateMachine/MetronomeStateMachineInterfaces";
+import {
+	DEFAULT_TEMPO,
+	E_DISPLAY_BLINKING_DELAYS,
+} from "@config/metronomeConfig";
+import { EStep } from "@config/commonInterfaces";
+
+export const INIT_METRONOME_STATE: IMetronomeContext = {
+	//TODO: change to DEFAULT_TEMPO
+	tempo: DEFAULT_TEMPO,
+	pattern: [EStep.HIGH, EStep.LOW, EStep.LOW, EStep.LOW],
+	isPlaying: false,
+	lastState: "tempoState",
+	tickTrigger: false,
+	display: {
+		text: DEFAULT_TEMPO.toString(),
+		blinkingChars: [],
+		blinkingDelay: E_DISPLAY_BLINKING_DELAYS.LONG,
+	},
+};
+
+export const SINGLE_TIMER_MACHINE_ID = "SINGLE_TIMER_MACHINE_ID";
