+++ src/services/MetronomeStateMachine/machines/MetronomeStateMachine/states/patternState/patternStateInterfaces.ts
@@ -1,0 +1,5 @@
+import { IMetronomeContext } from "@services/MetronomeStateMachine/machines/MetronomeStateMachine/MetronomeStateMachineInterfaces";
+
+export interface IPatternStateContext extends IMetronomeContext {
+	currentEditCharIndex: number;
+}
