+++ src/services/MetronomeStateMachine/machines/TimerStateMachine/TimerStateMachineConsts.ts
@@ -1,0 +1,6 @@
+import { DEFAULT_TEMPO } from "@config/metronomeConfig";
+
+export const DEFAULT_TIMER_CONTEXT = {
+	tempo: DEFAULT_TEMPO,
+	timeIntervalId: null,
+};
