+++ src/services/MetronomeStateMachine/machines/TimerStateMachine/states/timerPlayingState/timerPlayingStateInterfaces.ts
@@ -1,0 +1,16 @@
+import { ETimerStateMachineEventType } from "@services/MetronomeStateMachine/machines/TimerStateMachine/TimerStateMachineInterfaces";
+
+import { EStep } from "@config/commonInterfaces";
+
+export interface ITimerClearIntervalEvent {
+	type: ETimerStateMachineEventType.CLEAR_INTERVAL;
+}
+
+export type TTimerPlayingStateEvent = ITimerClearIntervalEvent;
+
+export type onTickHandler = (step: EStep) => void;
+
+export interface ITimerIntervalProps {
+	onTickHandler?: onTickHandler | onTickHandler[];
+	tempo: number;
+}
