--- src/services/MetronomeStateMachine/states/patternState/patternStateInterfaces.ts
@@ -1,5 +1,0 @@
-import { IMetronomeContext } from "@services/MetronomeStateMachine/MetronomeStateMachineInterfaces";
-
-export interface IPatternStateContext extends IMetronomeContext {
-	currentEditCharIndex: number;
-}
